Tighten TypeScript typings and remove casts
Clean up and tighten TypeScript typings across the codebase: remove unnecessary non-null assertions and redundant `as` casts, add a few explicit type casts where needed, and correct index/array typings. Notable changes include bech32 word typing, structuredClone generic cast, removing forced Bytes32/Script casts from crypto and payment modules, small witness/pubkey equality and indexing fixes, exporting SigningPoolLike, and a type-safe navigator check and WorkerSigningPool.getInstance call in worker code. These are type-safety and code-cleanup tweaks with no behavioral changes.

+++ src/psbt.ts
@@ -1049,7 +1049,7 @@ export class Psbt {
             sighashTypes,
         );
 
-        const sig = keyPair.sign(hash) as Uint8Array;
+        const sig = keyPair.sign(hash);
         const partialSig = [
             {
                 pubkey,
